refactor(db): move search semantics tests to crates/db/tests/
WHY: Batch G cluster 1 of 3. Tests that exercise search behavior without
trigger-side assertions land in search_semantics.rs. Helpers consumed
from common/mod.rs (Task 2). Test count unchanged (118 in-crate tests;
11 tests now in separate search_semantics integration-test binary).
Also fixes common/mod.rs: adds #![allow(unreachable_pub)] to suppress
clippy -D warnings for pub helpers in test-binary-local mod.
