+++ crates/db/tests/common/mod.rs
@@ -1,11 +1,18 @@
 #![allow(clippy::unwrap_used)] // WHY: integration test helpers; unwrap panics signal bugs.
-#![allow(dead_code)] // WHY: helpers are consumed by 3 sibling integration-test
-                     // binaries (search_semantics, search_triggers, search_proptests);
-                     // each binary compiles common/mod.rs but only uses a subset.
-                     // Without this, `dead_code` (workspace -D warnings) fires per
-                     // binary for helpers used elsewhere. Workaround per
-                     // rust-lang/rust#46379. Keep it permanent — adding a 4th test
-                     // binary later means the same problem recurs.
+#![allow(unreachable_pub)]
+// WHY: `pub` items in a test-binary-local `mod common` are
+// "unreachable" from rustc's perspective (no lib crate to
+// re-export them), but they ARE needed — each sibling binary
+// declares `mod common;` and uses a subset. Suppressed globally
+// so new helpers don't require per-item `#[allow]`.
+#![allow(dead_code)]
+// WHY: helpers are consumed by 3 sibling integration-test
+// binaries (search_semantics, search_triggers, search_proptests);
+// each binary compiles common/mod.rs but only uses a subset.
+// Without this, `dead_code` (workspace -D warnings) fires per
+// binary for helpers used elsewhere. Workaround per
+// rust-lang/rust#46379. Keep it permanent — adding a 4th test
+// binary later means the same problem recurs.
 
 //! Shared raw-SQL helpers for `crates/db/tests/search_*.rs` integration
 //! tests. Extracted from `crates/db/src/search_repo.rs::tests` in Batch G
@@ -17,15 +24,15 @@
 use std::path::Path;
 use std::sync::Arc;
 
-use perima_core::{DeviceId, EventBus};
+use perima_core::{DeviceId, EventBus, SearchRepository};
 use rusqlite::Connection;
 use tempfile::TempDir;
 
+use perima_db::SqliteSearchRepository;
 use perima_db::pool::ReadPool;
 use perima_db::tag_repo::SqliteTagRepository;
 use perima_db::test_utils::NoopBus;
 use perima_db::writer::{SqliteWriter, SqliteWriterHandle};
-use perima_db::SqliteSearchRepository;
 
 // ---------------------------------------------------------------------------
 // Cross-cluster constants
